test(vfs): split the monolithic vfs_e2e.rs into tests/vfs/*.rs by subsystem
The 1.9k-line tests/vfs_e2e.rs was unwieldy. Split it into one test binary
(tests/vfs.rs) that pulls in per-subsystem modules from tests/vfs/ via #[path]
(a crate-root `mod foo;` would resolve to tests/foo.rs, hence the explicit path):

- common.rs    — shared helpers/config (MODEL, s3/vfs/notion/gdrive/all configs,
                 provider, stamp, drive, verify_s3, task_for, tail) + re-exports.
- providers.rs — direct provider-adapter smoke (S3/Notion/GDrive, no frontend) [4]
- host.rs      — host FUSE frontend: non-sandbox mount, host-bind probe, inspect [3]
- sandbox.rs   — sandbox VM-restart lifecycle: remount/reconnect/concurrency/
                 multimount/agent-driven/inspect [8]
- forwarder.rs — the in-guest forwarder binary: crash recovery + static r/w/unlink [5]
- bench.rs     — host-vs-guest overhead benchmarks (+ shell_quote) [2]

Criterion: which part of the VFS stack each test exercises (core adapters /
host frontend / sandbox frontend / forwarder binary / benchmarks). Duplicated
helpers are centralized in common.rs and shared via `use crate::common::*`.
Pure move — all 22 tests preserved (verified via --list); the binary compiles
with no warnings and a sample test passes through the new layout.
